+++ examples/reference_tasks/spark_env.py
@@ -1,3 +1,7 @@
+"""
+This example demonstrates how to use the Spark environment in Flyte.
+"""
+
 from dataclasses import dataclass
 
 import flyte
