+++ examples/reference_tasks/torch_env.py
@@ -1,3 +1,7 @@
+"""
+Example that simulates a Torch environment using Flyte's TaskEnvironment.
+"""
+
 from pydantic import BaseModel
 
 import flyte
