+++ src/flyte/_bin/runtime.py
@@ -26,7 +26,6 @@ DOMAIN_NAME = "FLYTE_INTERNAL_TASK_DOMAIN"
 ORG_NAME = "_U_ORG_NAME"
 ENDPOINT_OVERRIDE = "_U_EP_OVERRIDE"
 RUN_OUTPUT_BASE_DIR = "_U_RUN_BASE"
-ENABLE_REF_TASKS = "_REF_TASKS"  # This is a temporary flag to enable reference tasks in the runtime.
 
 # TODO: Remove this after proper auth is implemented
 _UNION_EAGER_API_KEY_ENV_VAR = "_UNION_EAGER_API_KEY"
@@ -87,6 +86,7 @@ def main(
 
     import flyte
     import flyte._utils as utils
+    import flyte.errors
     from flyte._initialize import init
     from flyte._internal.controllers import create_controller
     from flyte._internal.imagebuild.image_builder import ImageCache
@@ -123,22 +123,7 @@ def main(
         logger.debug(f"Using controller endpoint: {ep} with kwargs: {controller_kwargs}")
 
     bundle = CodeBundle(tgz=tgz, pkl=pkl, destination=dest, computed_version=version)
-    enable_ref_tasks = os.getenv(ENABLE_REF_TASKS, "false").lower() in ("true", "1", "yes")
-    # We init regular client here so that reference tasks can work
-    # Current reference tasks will not work with remote controller, because we create 2 different
-    # channels on different threads and this is not supported by grpcio or the auth system. It ends up leading
-    # File "src/python/grpcio/grpc/_cython/_cygrpc/aio/completion_queue.pyx.pxi", line 147,
-    # in grpc._cython.cygrpc.PollerCompletionQueue._handle_events
-    # BlockingIOError: [Errno 11] Resource temporarily unavailable
-    # TODO solution is to use a single channel for both controller and reference tasks, but this requires a refactor
-    if enable_ref_tasks:
-        logger.warning(
-            "Reference tasks are enabled. This will initialize client and you will see a BlockIOError. "
-            "This is harmless, but a nuisance. We are working on a fix."
-        )
-        init(org=org, project=project, domain=domain, **controller_kwargs)
-    else:
-        init()
+    init(org=org, project=project, domain=domain, **controller_kwargs)
     # Controller is created with the same kwargs as init, so that it can be used to run tasks
     controller = create_controller(ct="remote", **controller_kwargs)
 
@@ -164,6 +149,8 @@ def main(
 
     # Run both coroutines concurrently and wait for first to finish and cancel the other
     async def _run_and_stop():
+        loop = asyncio.get_event_loop()
+        loop.set_exception_handler(flyte.errors.silence_grpc_polling_error)
         await utils.run_coros(controller_failure, task_coroutine)
         await controller.stop()
