+++ src/flyte/errors.py
@@ -10,6 +10,16 @@ from typing import Literal
 ErrorKind = Literal["system", "unknown", "user"]
 
 
+def silence_grpc_polling_error(loop, context):
+    """
+    Suppress specific gRPC polling errors in the event loop.
+    """
+    exc = context.get("exception")
+    if isinstance(exc, BlockingIOError):
+        return  # suppress
+    loop.default_exception_handler(context)
+
+
 class BaseRuntimeError(RuntimeError):
     """
     Base class for all Union runtime errors. These errors are raised when the underlying task execution fails, either
